requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.60.0

# Transformer & VLM Support (Qwen2.5-VL etc.)
transformers>=4.51.3
accelerate>=0.26.0
qwen-vl-utils>=0.0.4

# UniPercept-Reward (Aesthetic & Structural Scoring)
unipercept-reward>=0.1.0  # Install via pip or source

# Image Quality Assessment
pyiqa>=0.1.7

# Semantic Scoring (HPSv3)
# hpsv3  # Install from source or pip if available

# LLM API (OpenAI-compatible)
openai>=1.0.0

# Human Preference UI
gradio>=4.0.0

# MOS Calibration (Thurstone / Bradley-Terry)
choix>=0.3.5
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0
